add record batch struct
